fix: all allcoation of memory + add log + position of the new player defined by the egg position or the size of the map if no egg

+++ zappy_server/include/teams.h
@@ -22,14 +22,14 @@ enum orientation {
 
 typedef struct egg_s {
     int egg_id;
-    size_t x;
-    size_t y;
+    int x;
+    int y;
 } egg_t;
 
 typedef struct player_s {
     int id;
-    size_t x;
-    size_t y;
+    int x;
+    int y;
     size_t orientation;
     size_t level;
     size_t life;
